myproject/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
PyMySQL>=1.1
mysqlclient>=2.2
gunicorn>=21.0
dj-database-url>=2.0
whitenoise>=6.0
psycopg2-binary>=2.9